Forward hidden toast messages to host frame
